Included the sub id provided by the server
To identify the user is now retrieved the sub value
returned by the server. The string "_at_egi_unity" is
appended and stored as OpenID value of the user.

NOTE: there is build problem because a dependencies was updated
and the new version is not compatible. In order to compile you have
to fix the problem manually until a new release of the library
will be provided.